+++ tests/conftest.py
@@ -6,6 +6,7 @@ from substra.sdk.schemas import AlgoSpec
 from . import data_factory
 from .fl_interface import FLAlgoInputs
 from .fl_interface import FLAlgoOutputs
+from .fl_interface import AlgoCategory
 
 
 def pytest_configure(config):
@@ -79,11 +80,10 @@ def algo_query(tmpdir):
     desc_path = tmpdir / "description.md"
     desc_path.write_text("#Hello world", encoding="utf-8")
 
-    algo_category = substra.sdk.schemas.AlgoCategory.simple
+    algo_category = AlgoCategory.simple
 
     return AlgoSpec(
         name="algo_name",
-        category=algo_category,
         inputs=FLAlgoInputs[algo_category],
         outputs=FLAlgoOutputs[algo_category],
         description=str(desc_path),