+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- BREAKING CHANGE: remove category from substra.schema.AlgoSpec and substra.models.Algo
+
 ### Added
 
 - Prevent use of `__` in asset metadata keys in local mode